+++ src/test/java/org/patryk3211/electricity/TransmissionLineTest.java
@@ -53,7 +53,7 @@ public class TransmissionLineTest {
         Net2.W(R2, N2, null);
 
         for(int i = 0; i < 20; ++i) {
-            float Iprev = V1.getCurrent(), Vprev = N2.getVoltage();
+            double Iprev = V1.getCurrent(), Vprev = N2.getVoltage();
             Net1.calculate();
             Net2.calculate();
